fix insert overwrite static will fail cover more unit test
