+++ spark/v3.1/spark/src/main/java/com/netease/arctic/spark/io/TaskWriters.java
@@ -49,10 +49,12 @@ public class TaskWriters {
   private long taskId = 0;
   private StructType dsSchema;
 
+  private String unKeyedTmpDir;
   private final boolean isHiveTable;
   private final FileFormat fileFormat;
   private final long fileSize;
   private final long mask;
+
 
   protected TaskWriters(ArcticTable table) {
     this.table = table;
@@ -91,6 +93,11 @@ public class TaskWriters {
     return this;
   }
 
+  public TaskWriters withUnKeyedTmpDir(String unKeyedTmpDir) {
+    this.unKeyedTmpDir = unKeyedTmpDir;
+    return this;
+  }
+
   public TaskWriter<InternalRow> newBaseWriter(boolean isOverwrite) {
     preconditions();
 
@@ -124,7 +131,7 @@ public class TaskWriters {
     if (isHiveTable && isOverwrite) {
       outputFileFactory = new AdaptHiveOutputFileFactory(
           ((SupportHive) table).hiveLocation(), table.spec(), fileFormat, table.io(),
-          encryptionManager, partitionId, taskId, transactionId);
+          encryptionManager, partitionId, taskId, transactionId, unKeyedTmpDir);
     } else {
       outputFileFactory = new CommonOutputFileFactory(
           baseLocation, table.spec(), fileFormat, table.io(),
