+++ spark/v3.1/spark/src/main/java/com/netease/arctic/spark/writer/UnkeyedSparkBatchWrite.java
@@ -18,6 +18,7 @@
 
 package com.netease.arctic.spark.writer;
 
+import com.netease.arctic.hive.utils.HiveTableUtil;
 import com.netease.arctic.spark.io.TaskWriters;
 import com.netease.arctic.table.UnkeyedTable;
 import org.apache.iceberg.AppendFiles;
@@ -53,6 +54,7 @@ public class UnkeyedSparkBatchWrite implements ArcticSparkWriteBuilder.ArcticWri
 
   private final UnkeyedTable table;
   private final StructType dsSchema;
+  private final String unKeyedTmpDir = HiveTableUtil.getRandomSubDir();
 
   public UnkeyedSparkBatchWrite(UnkeyedTable table, StructType dsSchema) {
     this.table = table;
@@ -102,7 +104,7 @@ public class UnkeyedSparkBatchWrite implements ArcticSparkWriteBuilder.ArcticWri
 
     @Override
     public DataWriterFactory createBatchWriterFactory(PhysicalWriteInfo info) {
-      return new WriterFactory(table, dsSchema, false);
+      return new WriterFactory(table, dsSchema, false, unKeyedTmpDir);
     }
 
     @Override
@@ -119,7 +121,7 @@ public class UnkeyedSparkBatchWrite implements ArcticSparkWriteBuilder.ArcticWri
 
     @Override
     public DataWriterFactory createBatchWriterFactory(PhysicalWriteInfo info) {
-      return new WriterFactory(table, dsSchema, true);
+      return new WriterFactory(table, dsSchema, true, unKeyedTmpDir);
     }
 
     @Override
@@ -141,7 +143,7 @@ public class UnkeyedSparkBatchWrite implements ArcticSparkWriteBuilder.ArcticWri
 
     @Override
     public DataWriterFactory createBatchWriterFactory(PhysicalWriteInfo info) {
-      return new WriterFactory(table, dsSchema, true);
+      return new WriterFactory(table, dsSchema, true, unKeyedTmpDir);
     }
 
     @Override
@@ -149,6 +151,7 @@ public class UnkeyedSparkBatchWrite implements ArcticSparkWriteBuilder.ArcticWri
       OverwriteFiles overwriteFiles = table.newOverwrite();
       overwriteFiles.overwriteByRowFilter(overwriteExpr);
       for (DataFile file : files(messages)) {
+        System.out.println("------------data file ------------- " + file);
         overwriteFiles.addFile(file);
       }
       overwriteFiles.commit();
@@ -158,7 +161,7 @@ public class UnkeyedSparkBatchWrite implements ArcticSparkWriteBuilder.ArcticWri
   private class UpsertWrite extends BaseBatchWrite {
     @Override
     public DataWriterFactory createBatchWriterFactory(PhysicalWriteInfo info) {
-      return new DeltaUpsertWriteFactory(table, dsSchema);
+      return new DeltaUpsertWriteFactory(table, dsSchema, unKeyedTmpDir);
     }
 
     @Override
@@ -172,12 +175,16 @@ public class UnkeyedSparkBatchWrite implements ArcticSparkWriteBuilder.ArcticWri
     private final UnkeyedTable table;
     private final StructType dsSchema;
 
+    private final String unKeyedTmpDir;
+
     private boolean isOverwrite;
 
-    WriterFactory(UnkeyedTable table, StructType dsSchema, boolean isOverwrite) {
+
+    WriterFactory(UnkeyedTable table, StructType dsSchema, boolean isOverwrite, String unKeyedTmpDir) {
       this.table = table;
       this.dsSchema = dsSchema;
       this.isOverwrite = isOverwrite;
+      this.unKeyedTmpDir = unKeyedTmpDir;
     }
 
     @Override
@@ -186,6 +193,7 @@ public class UnkeyedSparkBatchWrite implements ArcticSparkWriteBuilder.ArcticWri
           .withPartitionId(partitionId)
           .withTaskId(taskId)
           .withDataSourceSchema(dsSchema)
+          .withUnKeyedTmpDir(unKeyedTmpDir)
           .newBaseWriter(this.isOverwrite);
       return new SimpleInternalRowDataWriter(writer);
     }
@@ -193,8 +201,8 @@ public class UnkeyedSparkBatchWrite implements ArcticSparkWriteBuilder.ArcticWri
 
   private static class DeltaUpsertWriteFactory extends WriterFactory {
 
-    DeltaUpsertWriteFactory(UnkeyedTable table, StructType dsSchema) {
-      super(table, dsSchema, false);
+    DeltaUpsertWriteFactory(UnkeyedTable table, StructType dsSchema, String unKeyedTmpDir) {
+      super(table, dsSchema, false, unKeyedTmpDir);
     }
 
     @Override
