+++ spark/v3.1/spark/src/test/java/com/netease/arctic/spark/SparkTestContext.java
@@ -185,6 +185,9 @@ public class SparkTestContext extends ExternalResource {
 
     sparkConfigs.put(SQLConf.PARTITION_OVERWRITE_MODE().key(), "DYNAMIC");
     sparkConfigs.put("spark.executor.heartbeatInterval", "300s");
+    sparkConfigs.put("spark.cores.max", "6");
+    sparkConfigs.put("spark.executor.cores", "2");
+    sparkConfigs.put("spark.default.parallelism", "12");
     sparkConfigs.put("spark.network.timeout", "500s");
     sparkConfigs.put("spark.sql.warehouse.dir", testSparkDir.getAbsolutePath());
     sparkConfigs.put("spark.sql.extensions", ArcticSparkExtensions.class.getName());
