+++ spark/v3.1/spark/src/test/java/com/netease/arctic/spark/hive/TestUnkeyedHiveInsertOverwriteDynamic.java
@@ -16,8 +16,9 @@
  * limitations under the License.
  */
 
-package com.netease.arctic.spark;
+package com.netease.arctic.spark.hive;
 
+import com.netease.arctic.spark.SparkTestBase;
 import org.apache.hadoop.hive.metastore.api.Partition;
 import org.apache.thrift.TException;
 import org.junit.After;
@@ -28,7 +29,7 @@ import org.junit.Test;
 import java.io.IOException;
 import java.util.List;
 
-public class TestUnkeyedHiveInsertOverwriteStatic extends SparkTestBase {
+public class TestUnkeyedHiveInsertOverwriteDynamic extends SparkTestBase {
 
   private final String database = "db";
   private final String table = "testA";
@@ -37,10 +38,6 @@ public class TestUnkeyedHiveInsertOverwriteStatic extends SparkTestBase {
 
   @Before
   public void before() throws IOException {
-    contextOverwriteMode = spark.conf().get("spark.sql.sources.partitionOverwriteMode");
-    System.out.println("spark.sql.sources.partitionOverwriteMode = " + contextOverwriteMode);
-    sql("set spark.sql.sources.partitionOverwriteMode = {0}", "STATIC");
-
     sql("use " + catalogNameHive);
     sql("create database if not exists {0}", database);
     sql("create table {0}.{1} ( \n" +
@@ -50,12 +47,17 @@ public class TestUnkeyedHiveInsertOverwriteStatic extends SparkTestBase {
         ") using arctic \n" +
         " partitioned by ( dt ) \n", database, table);
 
-
     sql("insert overwrite {0}.{1} values \n" +
         "(1, ''aaa'',  ''2021-1-1''), \n " +
         "(2, ''bbb'',  ''2021-1-2''), \n " +
         "(3, ''ccc'',  ''2021-1-3'') \n ", database, table);
 
+
+    sql("select * from {0}.{1} order by id", database, table);
+
+    contextOverwriteMode = spark.conf().get("spark.sql.sources.partitionOverwriteMode");
+    System.out.println("spark.sql.sources.partitionOverwriteMode = " + contextOverwriteMode);
+    sql("set spark.sql.sources.partitionOverwriteMode = {0}", "DYNAMIC");
   }
 
   @After
@@ -66,51 +68,47 @@ public class TestUnkeyedHiveInsertOverwriteStatic extends SparkTestBase {
   }
 
   @Test
-  public void testInsertOverwriteAllPartitionByValue() throws TException {
-    // insert overwrite by values, no partition expr
+  public void testInsertOverwrite() throws TException {
     sql("insert overwrite {0}.{1} values \n" +
         "(4, ''aaa'',  ''2021-1-1''), \n " +
         "(5, ''bbb'',  ''2021-1-2''), \n " +
         "(6, ''ccc'',  ''2021-1-2'') \n ", database, table);
 
-    rows = sql("select * from {0}.{1}", database, table);
-    Assert.assertEquals(3, rows.size());
-    assertContainIdSet(rows, 0, 4, 5, 6);
-
-    List<Partition> partitions = hms.getClient().listPartitions(
-        database,
-        table,
-        (short) -1);
-    Assert.assertEquals(2, partitions.size());
-
-    sql("use spark_catalog");
-    rows = sql("select * from {0}.{1}", database, table);
-    Assert.assertEquals(3, rows.size());
-    assertContainIdSet(rows, 0, 4, 5, 6);
-  }
-
-  @Test
-  public void testInsertOverwriteSomePartitionByValue() throws TException {
-    sql("insert overwrite {0}.{1} \n" +
-        "partition( dt = ''2021-1-1'')  values \n" +
-        "(4, ''aaa''), \n " +
-        "(5, ''bbb''), \n " +
-        "(6, ''ccc'') \n ", database, table);
-
-    rows = sql("select * from {0}.{1} order by id", database, table);
-    Assert.assertEquals(5, rows.size());
-    assertContainIdSet(rows, 0, 4, 5, 6, 2, 3);
+    rows = sql("select id, data, dt from {0}.{1} order by id", database, table);
+    Assert.assertEquals(4, rows.size());
+    assertContainIdSet(rows, 0, 4, 5, 6, 3);
 
     List<Partition> partitions = hms.getClient().listPartitions(
         database,
         table,
         (short) -1);
     Assert.assertEquals(3, partitions.size());
-
     sql("use spark_catalog");
-    rows = sql("select * from {0}.{1} order by id", database, table);
-    Assert.assertEquals(5, rows.size());
-    assertContainIdSet(rows, 0, 4, 5, 6, 2, 3);
+    rows = sql("select id, data, dt from {0}.{1} order by id", database, table);
+    Assert.assertEquals(4, rows.size());
+    assertContainIdSet(rows, 0, 4, 5, 6, 3);
+  }
+
+  @Test
+  public void testInsertOverwriteNoBasePartition() throws TException {
+    sql("insert overwrite {0}.{1} values \n" +
+        "(4, ''aaa'',  ''2021-1-4''), \n " +
+        "(5, ''bbb'',  ''2021-1-4''), \n " +
+        "(6, ''ccc'',  ''2021-1-4'') \n ", database, table);
+
+    rows = sql("select id, data, dt from {0}.{1} order by id", database, table);
+    Assert.assertEquals(6, rows.size());
+    assertContainIdSet(rows, 0, 1, 2, 3, 4, 5, 6);
+
+    List<Partition> partitions = hms.getClient().listPartitions(
+        database,
+        table,
+        (short) -1);
+    Assert.assertEquals(4, partitions.size());
+    sql("use spark_catalog");
+    rows = sql("select id, data, dt from {0}.{1} order by id", database, table);
+    Assert.assertEquals(6, rows.size());
+    assertContainIdSet(rows, 0, 1, 2, 3, 4, 5, 6);
   }
 
 }
